move: signature 생성 관련 코드를 util 폴더로 이동
